% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_search.R
\name{ann_search}
\alias{ann_search}
\title{Ann search function}
\usage{
ann_search(
  data,
  ll_set,
  search,
  grid = NULL,
  ll_object = NULL,
  threshold = 0.75
)
}
\arguments{
\item{data}{Data set.}

\item{ll_set}{Local lasso settings produced by \code{ll_set} function.}

\item{search}{"first" or "second".}

\item{grid}{Query points to find k-ANN.}

\item{ll_object}{Global variable relevancy produced by \code{local_lasso} function.}

\item{threshold}{Threshold defining variable relevancy (default = 0.75).}
}
\value{
k-ANN observations number and euclidean distance.
}
\description{
Ann search function
}
\details{
Find the k-ANN for query points.
}
