% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanechnikov.R
\name{epanechnikov}
\alias{epanechnikov}
\title{epanechnikov function}
\usage{
epanechnikov(u)
}
\arguments{
\item{u}{numerical variable.}
}
\value{
epanechnikov function at point u
}
\description{
epanechnikov function
}
\details{
Evaluate the value of epanechnikov function at point u.
}
