% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_grid.R
\name{gen_grid}
\alias{gen_grid}
\title{Generate grid function}
\usage{
gen_grid(ng, ll_set, x_var1, x_var2)
}
\arguments{
\item{ng}{Sequential length of \code{x_var1} and \code{x_var2} variables. The resulting grid has \code{ng*ng} rows.}

\item{ll_set}{Object produced by \code{ll_setting} function.}

\item{x_var1}{Name of the first variable of interest.}

\item{x_var2}{Name of the second variable of interest.}
}
\value{
Generated grid.
}
\description{
Generate grid function
}
\details{
Generate a grid of length \code{ng*ng}. Where variables of interest \code{x_var1} and \code{x_var2} variables
have a sequence of \code{ng} values and the remaining continuous variables are set to 0. The categorical variable are set to
one of the level observed.
}
