% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.plot.R
\name{ll.plot}
\alias{ll.plot}
\title{Default \code{ll} plotting}
\usage{
ll.plot(ll.object, x.var1, x.var2, ngrid, threshold = 0.75)
}
\arguments{
\item{ll.object}{\code{ll} object}

\item{x.var1}{variable to be ploted}

\item{x.var2}{variable to be ploted}

\item{ngrid}{number of grid values for each variable}

\item{threshold}{threshold to define global variables (default = 0.75)}
}
\value{
None
}
\description{
Default \code{ll} plotting
}
\details{
Produces default plot of conditional distribution of the response variable given covariates
x.var1 and x.var2.

The response variable is predicted for an expanded grid with ngrid*ngrid rows containing all combinations
of ngrid value of covariates x.var1 and x.var2. All others numerical covariates entries are set to 0 and
categorical covariates' set to random factor's level.

The threshold of global relevence is set by default to 0.75.
}
\examples{

#'library(locallasso)
set.seed(1993)

# # - - - - - - - - - - - - - - -#
# # Numerical Example
# # - - - - - - - - - - - - - - -#

# Define the data generating process
n=1000000   # number of observations
nc=20       # number of continuous variables

# Continous variables
X = mvrnorm(n,rep(0,nc),diag(1,nc))

# Categorical variables
D1 = factor(as.numeric(runif(n)<=.5))
D2 = factor(as.numeric(runif(n)<=.3))
F1 = factor(sample(LETTERS[1:4], n, replace=TRUE, prob=c(0.3, 0.3, 0.3, 0.1)))

# Response variable
y <- -X[,1]^2 - 4*sin((pi/2)*X[,2]) + X[,3]*X[,4] + rnorm(n,0,1)

data = data.frame(D1,X,y,D2,F1)

# Apply local lasso method to the data frame
ll = local.lasso(data,"y", k=500, nrand = 100)

# Plot of the conditional distribution of y given covariates X1 and X2.
ll.plot(ll,"X1","X2",ngrid=20,threshold = .75)

}
