% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.summary.R
\name{ll.summary}
\alias{ll.summary}
\title{Default \code{ll} summary}
\usage{
ll.summary(ll.object)
}
\arguments{
\item{ll.object}{\code{ll} object}
}
\value{
Summary of \code{ll} object
}
\description{
Default \code{ll} summary
}
\details{
Summary of results obtained by local lasso method.
}
\examples{

library(locallasso)
set.seed(1993)

# # - - - - - - - - - - - - - - -#
# # Numerical Example
# # - - - - - - - - - - - - - - -#

# Define the data generating process
n=1000000   # number of observations
nc=20       # number of continuous variables

# Continous variables
X = mvrnorm(n,rep(0,nc),diag(1,nc))

# Categorical variables
D1 = factor(as.numeric(runif(n)<=.5))
D2 = factor(as.numeric(runif(n)<=.3))
F1 = factor(sample(LETTERS[1:4], n, replace=TRUE, prob=c(0.3, 0.3, 0.3, 0.1)))

# Response variable
y <- -X[,1]^2 - 4*sin((pi/2)*X[,2]) + X[,3]*X[,4] + rnorm(n,0,1)

data = data.frame(D1,X,y,D2,F1)

# Apply local lasso method to the data frame
ll = local.lasso(data,"y", k=500, nrand = 100)

ll.summary(ll)
}
