% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_settings.R
\name{ll_settings}
\alias{ll_settings}
\title{Local lasso setting function}
\usage{
ll_settings(sample, response, k, e, eps = 1)
}
\arguments{
\item{sample}{Random sample drawn from full data set with \code{sampling} function.}

\item{response}{Name of the response variable.}

\item{k}{Number of approximate nearest neighbours.}

\item{e}{Number of evaluation points.}

\item{eps}{Value of \code{epsilon} for the ANN-search (default = 1).}
}
\value{
Settings for local lasso function.
}
\description{
Local lasso setting function
}
\details{
Compute a robust estimation of standard deviations. Generate random set of evaluation points.
}
