% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_summary.R
\name{ll_summary}
\alias{ll_summary}
\title{Local lasso summary function}
\usage{
ll_summary(ll_object)
}
\arguments{
\item{ll_object}{Local lasso object produced by the \code{local_lasso} function.}
}
\value{
Summary of local lasso estimation results.
}
\description{
Local lasso summary function
}
\details{
Display global relevancy scores and KS-statistics for testing uniformity of relevancy points.
}
