% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ll.R
\name{plot_ll}
\alias{plot_ll}
\title{Plot function}
\usage{
plot_ll(ann, grid, ll_object, x_var1, x_var2, threshold = 0.75)
}
\arguments{
\item{ann}{Approximate nearest neighbours obtained by \code{ann_search} function.}

\item{grid}{Grid points to evaluate produced by \code{gen_grid} function.}

\item{ll_object}{Local lasso object produced by the \code{local_lasso} function.}

\item{x_var1}{Name of the first variable of interest.}

\item{x_var2}{Name of the second variable of interest.}

\item{threshold}{Threshold defining variable relevancy (default = 0.75).}
}
\value{
Plot of the conditional expectation of response variable given \code{x_var1} and \code{x_var2} variables.
}
\description{
Plot function
}
\details{
Predict the value of response for \code{grid}.
Plot of the conditional expectation of response variable given \code{x_var1} and \code{x_var2} variables.
}
