% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ll.R
\name{pred_ll}
\alias{pred_ll}
\title{Predict function}
\usage{
pred_ll(ann, newdata, ll_object, threshold = 0.75)
}
\arguments{
\item{ann}{Approximate nearest neighbours obtained by \code{ann_search} function.}

\item{newdata}{Set of points of interest to predict}

\item{ll_object}{Local lasso object produced by the \code{local_lasso} function.}

\item{threshold}{Threshold defining variable relevancy (default = 0.75).}
}
\value{
Prediction of response variable for \code{newdata}
}
\description{
Predict function
}
\details{
Predict the value of response for new set of point provided.
}
