% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling}
\alias{sampling}
\title{Sampling function}
\usage{
sampling(data, ss = 1000, fixed = T, frac = 0.01)
}
\arguments{
\item{data}{Data set.}

\item{ss}{Number of observation to sample from full dataset (default = 1000)}

\item{fixed}{Sample a fixed number of observation (fixed = T) or a fraction of the data set (fixed = F)}

\item{frac}{Fraction of observation to sample from dataset (default = 0.01)}
}
\value{
Random sample of observation
}
\description{
Sampling function
}
\details{
Draw a random sample of observation from data set.
}
