% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanechnikov.R
\name{Epanechnikov}
\alias{Epanechnikov}
\title{Epanechnikov function}
\usage{
Epanechnikov(u)
}
\arguments{
\item{u}{numerical variable.}
}
\value{
Epanechnikov function at point u
}
\description{
Epanechnikov function
}
\details{
Evaluate the value of Epanechnikov function at point u.
}
